requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
dj-rest-auth==1.1.1
Django==3.0.8
django-braces==1.14.0
django-cors-headers==3.5.0
django-cors-middleware==1.5.0
django-heroku==0.3.1
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
